refactor(agent-teams): split oversized files and remove unused MCB fallback param
