+++ src/tools/agent-teams/team-member-utils.ts
@@ -1,0 +1,5 @@
+import type { TeamMember, TeamTeammateMember } from "./types"
+
+export function isTeammateMember(member: TeamMember): member is TeamTeammateMember {
+  return member.agentType !== "team-lead"
+}
